Send SIGINT to docker for quick shutdown
The --signal=2 flag corresponds to SIGINT, a signal the main.py
controller is listening for. When received it lowers the plate and
clears the OLED screen.

The default signal is SIGKILL wasn't being caught, resulting in the
controller quitting without lowering the plate.

Also, this service runs as "pi" since RPi.GPIO is installed with pip
install --user. This user *does* have password-less sudo privelages, as
part of default Debian install.