2.4.1 (#3)
* Send SIGINT to docker for quick shutdown

The --signal=2 flag corresponds to SIGINT, a signal the main.py
controller is listening for. When received it lowers the plate and
clears the OLED screen.

The default signal is SIGKILL wasn't being caught, resulting in the
controller quitting without lowering the plate.

Also, this service runs as "pi" since RPi.GPIO is installed with pip
install --user. This user *does* have password-less sudo privelages, as
part of default Debian install.

* Plate was slamming on some production bots

This lowers it to closer (155 degrees), but not all the way.

* New MOABIAN env variable will contain version num

Single point to update for new releases: moab/os/files/etc/environment.
It's copied to /etc/environment and sourced on login.

* Install pre-built library

Source is in moab/sw/libs/moab (and oled). By shipping it pre-built,
means compiler doesn't need