New: offset tool to add +/- values to each servo
Although the servo arms come adjusted and calibrated during assembly,
we've found that local level variations (i.e., your table isn't actually
level) or tolerance stacking requires small tweaks for your Moab.

It's helpful to think of the 3 servos as positions on a clock dial.  With the
joystick positioned nearest you, at the 6 o'clock position, the servos are:

1: 12 o'clock (top) position
2: 8 o'clock
3: 4 o'clock

Typical usage is a manual, iterative process:

```
% down                  # first stop moab controller

% offset                # show current offsets (from ~/moab/config/calibration.json)
% offset -s 5 0 0       # Say the 12 o'clock servo is lower than the others...test
% offset -s 4 0 1       # A little too high...this one is perfect.
% offset -s 4 0 1 -w    # Write the changes

% up            # restart controller with new offsets
```